Update mount-rosetta.py
I couldn't get it to load FUSE so I tried to load fuse module in python. There was no module named fuse. I didn't install anything named fuse, but I did install fusepy. So I updated line 11 by changing 'fuse' to 'fusepy' and it worked. That is ALL the work I did on this. 100% of the credit for this goes to the original creator of this repository, mrexodia. 

Original: https://github.com/mrexodia/rosetta-multipass/edit/main/mount-rosetta.py